Split up session-helper into session-core and session-helper
